distance cleanup
